Move HDULocation to utils/fits.py
